refactor(local-ci): extract git_helpers.py for git + time helpers
R2-1 third cut. Move 6 helpers (now_iso, current_branch, current_sha,
git_root_for, resolve_git_ref_sha, short_sha) plus the shared ROOT
constant out of local_ci.py into a sibling git_helpers.py module.

- local_ci.py: 11,078 → 11,033 lines (-45 net).
- git_helpers.py: 79 lines.
- ROOT is defined in both modules; resolution is identical because both
  files live at tools/local-ci/<file>.py.
- All 6 symbols re-export from local_ci.py for backward compat.

Tests: 213/213 local_ci tests pass.
