+++ tools/local-ci/local_ci.py
@@ -484,59 +484,14 @@ def file_lock(path: Path, *, blocking: bool):
         handle.close()
 
 
-def now_iso() -> str:
-    return datetime.now(timezone.utc).isoformat()
-
-
-def current_branch() -> str:
-    result = subprocess.run(
-        ["git", "rev-parse", "--abbrev-ref", "HEAD"],
-        cwd=ROOT,
-        capture_output=True,
-        text=True,
-        check=True,
-    )
-    return result.stdout.strip()
-
-
-def current_sha() -> str:
-    result = subprocess.run(
-        ["git", "rev-parse", "HEAD"],
-        cwd=ROOT,
-        capture_output=True,
-        text=True,
-        check=True,
-    )
-    return result.stdout.strip()
-
-
-def git_root_for(path: Path) -> Path | None:
-    result = subprocess.run(
-        ["git", "rev-parse", "--show-toplevel"],
-        cwd=path,
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        return None
-    return Path(result.stdout.strip()).resolve()
-
-
-def resolve_git_ref_sha(ref: str) -> str:
-    result = subprocess.run(
-        ["git", "rev-parse", "--verify", f"{ref}^{{commit}}"],
-        cwd=ROOT,
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        detail = (result.stderr or result.stdout or "").strip()
-        raise ValueError(f"Could not resolve git ref '{ref}': {detail or 'unknown ref'}")
-    return result.stdout.strip()
-
-
-def short_sha(sha: str) -> str:
-    return sha[:12] if sha else "?"
+from git_helpers import (  # noqa: E402  -- re-exported for in-file consumers
+    now_iso,
+    current_branch,
+    current_sha,
+    git_root_for,
+    resolve_git_ref_sha,
+    short_sha,
+)
 
 
 from normalize import (  # noqa: E402  -- re-exported for in-file consumers
